Promt for salesRep to create lead, create salesRep functionality, refactor opportunity menu, change enum types

+++ src/main/java/com/ironhack/project_crm_2/classes/menus/OpportunityMenu.java
@@ -32,23 +32,38 @@ public class OpportunityMenu {
         this.SALES_REP_SERVICE = salesRepService;
     }
 
-    public void convertLeadIntoOpportunity(){
-        Lead leadToConvert = LEAD_SERVICE.getLeadToConvert();
-        ContactInfo contactInfo = getInfoLeadToConvert(leadToConvert);
-        Account account = selectAccountForOpportunity();
-        contactInfo.account = account;
-        Contact decisionMaker = CONTACT_SERVICE.createContact(contactInfo);
-        OpportunityInfo opportunityInfo = getOpportunityInfo(decisionMaker, account);
-        Opportunity newOpportunity = OPPORTUNITY_SERVICE.createOpportunity(opportunityInfo);
-        LEAD_SERVICE.delete(leadToConvert.getId());
+    public void convertLeadIntoOpportunity() {
+        if (LEAD_SERVICE.isEmptyList()) {
+            System.err.println("Lead list is empty.");
+        } else {
+            Lead leadToConvert = getLeadToConvert();
+            SalesRep salesRep = leadToConvert.getSalesRep();
+            ContactInfo contactInfo = getInfoLeadToConvert(leadToConvert);
+            Account account = selectAccountForOpportunity();
+            contactInfo.account = account;
+            Contact decisionMaker = CONTACT_SERVICE.createContact(contactInfo);
+            OpportunityInfo opportunityInfo = getOpportunityInfo(decisionMaker, account, salesRep);
+            Opportunity newOpportunity = OPPORTUNITY_SERVICE.createOpportunity(opportunityInfo);
+            LEAD_SERVICE.delete(leadToConvert.getId());
+        }
     }
 
-    public OpportunityInfo getOpportunityInfo(Contact decisionMaker, Account account) {
-        ProductType productType = chooseProductType();
-        int numProducts = Utils.promptForInt("Number of products:");
+    public Lead getLeadToConvert() {
+            while (true) {
+                try {
+                    int id = Utils.promptForInt("Enter a valid lead ID: ");
+                    Lead leadToConvert = LEAD_SERVICE.getById(id);
+                    System.out.println(leadToConvert.toString());
+                    return leadToConvert;
+                } catch (IllegalArgumentException | InputMismatchException e) {
+                    System.err.println(e.getMessage());
+                }
+            }
+    }
 
-        //Pending manage error if sales rep not found
-        SalesRep salesRep = SALES_REP_SERVICE.getById(Utils.promptForInt("Select sales repr by Id"));
+    public OpportunityInfo getOpportunityInfo(Contact decisionMaker, Account account, SalesRep salesRep) {
+        ProductType productType = chooseProductType();
+        int numProducts = Utils.promptForInt("Number of products: ");
 
         return new OpportunityInfo(
                 productType,
@@ -60,6 +75,12 @@ public class OpportunityMenu {
             );
         }
 
+
+    public static void displayCreateOrAssociateAccountMenu() {
+        System.out.println("1. Create account");
+        System.out.println("2. Associate to existent account");
+    }
+
     public Account selectAccountForOpportunity(){
         displayCreateOrAssociateAccountMenu();
         String choice = INPUT.nextLine();
@@ -69,32 +90,33 @@ public class OpportunityMenu {
                     AccountInfo accountInfo = requestAccountInfo();
                     return ACCOUNT_SERVICE.createAccount(accountInfo);
                 case "2":  //Select account by id
-                    return requestAccountById();
+                    if(ACCOUNT_SERVICE.isEmptyList()) {
+                        System.err.println("Account list is empty.");
+                        break;
+                    } else return ACCOUNT_SERVICE.requestAccountById();
                 default:
                     System.err.println("Please select a valid option");
             }
         }
     }
 
-    public Account requestAccountById(){
-        while (true){
-            int id = Utils.promptForInt("Enter account Id");
-            try {
-                Account selectedAccount = ACCOUNT_SERVICE.getById(id);
-                return selectedAccount;
-            } catch (IllegalArgumentException | InputMismatchException e) {
-                System.err.println(e.getMessage());
-            }
-        }
-    }
 
     public AccountInfo requestAccountInfo() {
         return new AccountInfo(
                 requestIndustryOption(),
                 Utils.promptForInt("Number of employees: "),
                 Utils.promptForString("City: "),
-                Utils.promptForString("Number: ")
+                Utils.promptForString("Country: ")
         );
+    }
+
+    public static void displayIndustryOptionMenu(){
+        System.out.println("Choose the company's sector:");
+        System.out.println("1. Produce");
+        System.out.println("2. E-Commerce");
+        System.out.println("3. Manufacturing");
+        System.out.println("4. Medical");
+        System.out.println("5. Other");
     }
 
     public IndustryOption requestIndustryOption() {
@@ -112,8 +134,9 @@ public class OpportunityMenu {
                     return IndustryOption.MEDICAL;
                 case "5":
                     return IndustryOption.OTHER;
+                default:
+                    System.err.println("Please select a valid option");
             }
-            System.err.println("Please select a valid option");
         }
     }
 
@@ -127,6 +150,13 @@ public class OpportunityMenu {
                 null);
         }
 
+    public static void displayOpportunityTypeMenu() {
+        System.out.println("Choose product type:");
+        System.out.println("1. Hybrid");
+        System.out.println("2. Flatbed");
+        System.out.println("3. Box");
+    }
+
 
     public ProductType chooseProductType() {
         displayOpportunityTypeMenu();
@@ -139,29 +169,11 @@ public class OpportunityMenu {
                     return ProductType.FLATBED;
                 case "3":
                     return ProductType.BOX;
+                default:
+                    System.err.println("Please select a valid option");
             }
-            System.err.println("Please select a valid option");
         }
     }
 
-    public static void displayIndustryOptionMenu(){
-        System.out.println("Choose the company's sector:");
-        System.out.println("1. Produce");
-        System.out.println("2. E-Commerce");
-        System.out.println("3. Manufacturing");
-        System.out.println("4. Medical");
-        System.out.println("5. Other");
-    }
 
-    public static void displayOpportunityTypeMenu() {
-        System.out.println("Choose product type:");
-        System.out.println("1. Hybrid");
-        System.out.println("2. Flatbed");
-        System.out.println("3. Box");
-    }
-
-    public static void displayCreateOrAssociateAccountMenu() {
-        System.out.println("1. Create account");
-        System.out.println("2. Associate to existent account");
-    }
 }
